added routes for jwt authentication

+++ src/services/posts.js
@@ -1,7 +1,7 @@
 import { Post } from '../db/models/post.js';
 
-export async function createPost({ title, author, contents, tags }) {
-	const post = new Post({ title, author, contents, tags });
+export async function createPost(userId, { title, contents, tags }) {
+	const post = new Post({ title, author: userId, contents, tags });
 	return await post.save();
 }
 
@@ -28,9 +28,13 @@ export async function getPostById(postId) {
 	return await Post.findById(postId);
 }
 
-export async function updatePost(postId, { title, author, contents, tags }) {
+export async function updatePost(
+	userId,
+	postId,
+	{ title, author, contents, tags }
+) {
 	return await Post.findByIdAndUpdate(
-		{ _id: postId },
+		{ _id: postId, author: userId },
 		{
 			$set: {
 				title,
@@ -43,8 +47,6 @@ export async function updatePost(postId, { title, author, contents, tags }) {
 	);
 }
 
-export async function deletePost(postId) {
-	return await Post.deleteOne({ _id: postId });
+export async function deletePost(userId, postId) {
+	return await Post.deleteOne({ _id: postId, author: userId });
 }
-
-
